refactor: consolidate cell phone implementation
