+++ discord_bot/agent_interface.py
@@ -6,7 +6,7 @@ Provides an interface between Discord commands and the Dream.OS Cell Phone syste
 
 import logging
 from typing import Dict, List, Optional, Any
-from dreamos.core.cell_phone import CellPhone
+from dreamos.core.messaging.cell_phone import CellPhone
 from dreamos.core.messaging.types import MessageMode
 
 logger = logging.getLogger('discord_bot.agent_interface')
