+++ dreamos/core/message_processor.py
@@ -7,7 +7,7 @@ Handles message processing and routing between agents.
 import logging
 from typing import Dict, List, Optional, Any
 from datetime import datetime
-from .cell_phone import CellPhone
+from .messaging.cell_phone import CellPhone
 from .messaging.types import Message, MessageMode
 from .cursor_controller import CursorController
 from .coordinate_manager import CoordinateManager
