+++ tests/agent_control/test_devlog_manager.py
@@ -46,9 +46,13 @@ sys.modules.setdefault("discord.ext", mock_discord.ext)
 sys.modules.setdefault("discord.ext.commands", mock_discord.ext.commands)
 sys.modules.setdefault("discord.ext.tasks", mock_discord.ext.tasks)
 sys.modules.setdefault("aiohttp", types.ModuleType("aiohttp"))
-cell_phone_stub = types.ModuleType("dreamos.core.cell_phone")
-cell_phone_stub.CaptainPhone = object
-sys.modules.setdefault("dreamos.core.cell_phone", cell_phone_stub)
+cell_phone_stub = types.ModuleType("dreamos.core.messaging.cell_phone")
+cell_phone_stub.CaptainPhone = type("CaptainPhone", (), {"__init__": lambda *_, **__: None})
+cell_phone_stub.CellPhone = type("CellPhone", (), {"__init__": lambda *_, **__: None})
+sys.modules.setdefault("dreamos.core.messaging.cell_phone", cell_phone_stub)
+captain_phone_stub = types.ModuleType("dreamos.core.messaging.captain_phone")
+captain_phone_stub.CaptainPhone = type("CaptainPhone", (), {"__init__": lambda *_, **__: None})
+sys.modules.setdefault("dreamos.core.messaging.captain_phone", captain_phone_stub)
 message_stub = types.ModuleType("dreamos.core.messaging.message")
 message_stub.Message = object
 sys.modules.setdefault("dreamos.core.messaging.message", message_stub)
